add EventListerner checkbox
